Extracting cumulative code hook
