+++ src/components/code-cell.tsx
@@ -6,7 +6,7 @@ import Resizable from "./resizable";
 import { Cell } from "../state";
 import { useActions } from "../hooks/use-actions";
 import { useTypedSelector } from "../hooks/use-typed-selector";
-import { CellType } from "../enums";
+import { useCumulativeCode } from "../hooks/use-cumulative-code";
 
 interface CodeCellProps {
   cell: Cell;
@@ -16,55 +16,16 @@ const CodeCell: React.FC<CodeCellProps> = ({ cell }) => {
   //user will not be able to use local storage or cookies with this app
   const { updateCell, createBundle } = useActions();
   const bundle = useTypedSelector((state) => state.bundles[cell.id]);
-  const cumulativeCode = useTypedSelector((state) => {
-    const { data, order } = state.cells;
-    const orderedCells = order.map((id) => data[id]);
-
-    const showFunction = `
-      import _React from 'react';
-      import _ReactDOM from 'react-dom';
-      var show = (value) => {
-        const root = document.querySelector('#root');
-        if (typeof value === 'object') {
-          if (value.$$typeof && value.props) {
-            _ReactDOM.render(value, root);
-          } else {
-            root.innerHTML = JSON.stringify(value);
-          }
-        } else {
-          root.innerHTML = value
-        }
-      }
-      `;
-
-    const showFunctionNoOp = "var show = ()=> {}";
-
-    const cumulativeCode = [];
-
-    for (let c of orderedCells) {
-      if (c.type === CellType.CODE) {
-        if (c.id === cell.id) {
-          cumulativeCode.push(showFunction);
-        } else {
-          cumulativeCode.push(showFunctionNoOp);
-        }
-        cumulativeCode.push(c.content);
-      }
-      if (c.id === cell.id) {
-        break;
-      }
-    }
-    return cumulativeCode;
-  });
+  const cumulativeCode = useCumulativeCode(cell.id);
 
   useEffect(() => {
     if (!bundle) {
-      createBundle(cell.id, cumulativeCode.join("\n"));
+      createBundle(cell.id, cumulativeCode);
       return;
     }
     //setting up debouncer for bundler
     const timer = setTimeout(async () => {
-      createBundle(cell.id, cumulativeCode.join("\n"));
+      createBundle(cell.id, cumulativeCode);
     }, 750);
 
     //cleanup Fn for the timer
@@ -72,7 +33,7 @@ const CodeCell: React.FC<CodeCellProps> = ({ cell }) => {
       clearTimeout(timer);
     };
     // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [cumulativeCode.join("\n"), cell.id, createBundle]);
+  }, [cumulativeCode, cell.id, createBundle]);
 
   return (
     <Resizable direction="vertical">
